Include install instructions on manage page for cmd tokens
